components/optimization-engine/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
exceptiongroup==1.3.0
idna==3.10
kafka-python==2.2.13
multidict==6.5.0
networkx==3.5
numpy==2.3.1
pamqp==3.3.0
propcache==0.3.2
PyYAML==6.0.2
requests==2.32.4
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1